+++ src/main/resources/templates/fashionshop-ui/src/layouts/admin/authentication/sign-up/index.jsx
@@ -1,0 +1,104 @@
+
+import { useState } from "react";
+
+// react-router-dom components
+import { Link } from "react-router-dom";
+
+// @mui material components
+import Card from "@mui/material/Card";
+import Checkbox from "@mui/material/Checkbox";
+
+// Soft UI Dashboard React components
+import SoftBox from "../../../../components/SoftBox";
+import SoftTypography from "../../../../components/SoftTypography";
+import SoftInput from "../../../../components/SoftInput";
+import SoftButton from "../../../../components/SoftButton";
+
+// Authentication layout components
+import BasicLayout from "../components/BasicLayout";
+import Socials from "../components/Socials";
+import Separator from "../components/Separator";
+
+// Images
+import curved6 from "../../../../assets/images/curved-images/curved14.jpg";
+
+function SignUp() {
+  const [agreement, setAgremment] = useState(true);
+
+  const handleSetAgremment = () => setAgremment(!agreement);
+
+  return (
+    <BasicLayout
+      title="Welcome!"
+      description="Use these awesome forms to login or create new account in your project for free."
+      image={curved6}
+    >
+      <Card>
+        <SoftBox p={3} mb={1} textAlign="center">
+          <SoftTypography variant="h5" fontWeight="medium">
+            Register with
+          </SoftTypography>
+        </SoftBox>
+        <SoftBox mb={2}>
+          <Socials />
+        </SoftBox>
+        <Separator />
+        <SoftBox pt={2} pb={3} px={3}>
+          <SoftBox component="form" role="form">
+            <SoftBox mb={2}>
+              <SoftInput placeholder="Name" />
+            </SoftBox>
+            <SoftBox mb={2}>
+              <SoftInput type="email" placeholder="Email" />
+            </SoftBox>
+            <SoftBox mb={2}>
+              <SoftInput type="password" placeholder="Password" />
+            </SoftBox>
+            <SoftBox display="flex" alignItems="center">
+              <Checkbox checked={agreement} onChange={handleSetAgremment} />
+              <SoftTypography
+                variant="button"
+                fontWeight="regular"
+                onClick={handleSetAgremment}
+                sx={{ cursor: "poiner", userSelect: "none" }}
+              >
+                &nbsp;&nbsp;I agree the&nbsp;
+              </SoftTypography>
+              <SoftTypography
+                component="a"
+                href="#"
+                variant="button"
+                fontWeight="bold"
+                textGradient
+              >
+                Terms and Conditions
+              </SoftTypography>
+            </SoftBox>
+            <SoftBox mt={4} mb={1}>
+              <SoftButton variant="gradient" color="dark" fullWidth>
+                sign up
+              </SoftButton>
+            </SoftBox>
+            <SoftBox mt={3} textAlign="center">
+              <SoftTypography variant="button" color="text" fontWeight="regular">
+                Already have an account?&nbsp;
+                <SoftTypography
+                  component={Link}
+                  to="/authentication/sign-in"
+                  variant="button"
+                  color="dark"
+                  fontWeight="bold"
+                  textGradient
+                >
+                  Sign in
+                </SoftTypography>
+              </SoftTypography>
+            </SoftBox>
+          </SoftBox>
+        </SoftBox>
+      </Card>
+    </BasicLayout>
+  );
+}
+
+export default SignUp;
